+++ lib/src/services/tokener.dart
@@ -1,6 +1,5 @@
 import 'dart:convert';
 import 'dart:typed_data';
-
 import 'package:crypto/crypto.dart';
 import 'package:encrypt/encrypt.dart';
 
@@ -11,6 +10,20 @@ class ZeytinTokener {
   ZeytinTokener(String passphrase)
     : key = _deriveKey(passphrase),
       encrypter = Encrypter(AES(_deriveKey(passphrase), mode: AESMode.cbc));
+
+  String encryptString(String text) {
+    final iv = IV.fromSecureRandom(16);
+    final encrypted = encrypter.encrypt(text, iv: iv);
+    return "${iv.base64}:${encrypted.base64}";
+  }
+
+  String decryptString(String encryptedData) {
+    final parts = encryptedData.split(':');
+    if (parts.length != 2) throw FormatException("Invalid format");
+    final iv = IV.fromBase64(parts[0]);
+    final decrypted = encrypter.decrypt(Encrypted.fromBase64(parts[1]), iv: iv);
+    return decrypted;
+  }
 
   static Key _deriveKey(String passphrase) {
     final bytes = utf8.encode(passphrase);