+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 * TODO: Describe initial release.
+
+## 2.1.0
+
+* Major changes on the server side: The token acquisition mechanism has been completely changed; email and password no longer circulate within the payload.
+* Changes in system infrastructure.